Move order loading methods to list VM
